Replace preset tabs with tree-based resources browser
