+++ src/components/ResourcesModal.css
@@ -619,6 +619,58 @@
   }
 }
 
+/* Resources browser layout */
+.resources-layout {
+  flex: 1;
+  display: flex;
+  height: 100%;
+}
+
+.resources-tree {
+  width: var(--sidebar-width, 220px);
+  background: rgba(0, 0, 0, 0.2);
+  overflow-y: auto;
+  padding: 8px 0;
+}
+
+.resources-resizer {
+  width: 5px;
+  cursor: col-resize;
+  background: rgba(255, 255, 255, 0.1);
+}
+
+.resources-details {
+  flex: 1;
+  background: #1a1a1a;
+  overflow-y: auto;
+  padding: 20px;
+}
+
+.tree-node {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  padding: 2px 8px;
+  cursor: pointer;
+  user-select: none;
+  white-space: nowrap;
+}
+
+.tree-node.selected {
+  background: rgba(255, 255, 255, 0.1);
+}
+
+.tree-node .expander {
+  width: 12px;
+  display: inline-block;
+  text-align: center;
+}
+
+.tree-children {
+  margin-left: 12px;
+}
+
+
 @media (max-width: 768px) {
   .preset-gallery-content {
     flex-direction: column;
